优化图片视频展示和保存功能
